feat(Runner): add sync tasks
This new feature allows you to set up tasks that block the chain until
completed. This is useful when tasks modify the same file, which would
otherwise cause errors (race conditions).

BREAKING CHANGE: Prepare is now always run, even for definedTasks. When
returning anything of type object from prepare, it will replace the
parameters in the chain. This is useful for simpel prepare changed.