feat(Runner): add support for dynamicTask
Dynamic tasks allow you to define the rules from a callback. The cool
thing is that this can also be a promise, meaning async dynamic tasks
are now officially a thing! Make an API call, read a dir, query a
database or ask for user input on the CLI.

BREAKING CHANGE: Exceptions are no longer thrown. Instead, they're now
rejected promises. The reason for this is to normalize the return value
to a Promise and make the entire Runner.run more predictable.